requirements.txt
fastapi
uvicorn
fitz
requests
sentence-transformers
faiss-cpu
streamlit
pdfplumber